Move ObsoleteCallStatementInspectionTests for QuickFixes into relevant files under RD.QuickFixes
