+++ RubberduckTests/Inspections/ObsoleteCallStatementInspectionTests.cs
@@ -3,7 +3,6 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using RubberduckTests.Mocks;
 using System.Threading;
 using Rubberduck.Inspections.Concrete;
-using Rubberduck.Inspections.QuickFixes;
 using Rubberduck.Parsing.Inspections.Resources;
 using Rubberduck.VBEditor.SafeComWrappers.Abstract;
 
@@ -174,83 +173,6 @@ End Sub";
             Assert.IsFalse(inspectionResults.Any());
         }
 
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void ObsoleteCallStatement_QuickFixWorks_RemoveCallStatement()
-        {
-            const string inputCode =
-@"Sub Foo()
-    Call Goo(1, ""test"")
-End Sub
-
-Sub Goo(arg1 As Integer, arg1 As String)
-    Call Foo
-End Sub";
-
-            const string expectedCode =
-@"Sub Foo()
-    Goo 1, ""test""
-End Sub
-
-Sub Goo(arg1 As Integer, arg1 As String)
-    Foo
-End Sub";
-            
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out var component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new ObsoleteCallStatementInspection(state);
-            var inspector = InspectionsHelper.GetInspector(inspection);
-            var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
-
-            var fix = new RemoveExplicitCallStatmentQuickFix(state);
-            foreach (var result in inspectionResults)
-            {
-                fix.Fix(result);
-            }
-
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void ObsoleteCallStatement_IgnoreQuickFixWorks()
-        {
-            const string inputCode =
-@"Sub Foo()
-    Call Goo(1, ""test"")
-End Sub
-
-Sub Goo(arg1 As Integer, arg1 As String)
-    Call Foo
-End Sub";
-
-            const string expectedCode =
-@"Sub Foo()
-'@Ignore ObsoleteCallStatement
-    Call Goo(1, ""test"")
-End Sub
-
-Sub Goo(arg1 As Integer, arg1 As String)
-'@Ignore ObsoleteCallStatement
-    Call Foo
-End Sub";
-            
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out var component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new ObsoleteCallStatementInspection(state);
-            var inspector = InspectionsHelper.GetInspector(inspection);
-            var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
-
-            var fix = new IgnoreOnceQuickFix(state, new[] {inspection});
-            foreach (var result in inspectionResults)
-            {
-                fix.Fix(result);
-            }
-
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
 
         [TestMethod]
         [TestCategory("Inspections")]
