+++ RubberduckTests/QuickFixes/IgnoreOnceQuickFixTests.cs
@@ -585,5 +585,47 @@ End Sub";
             new IgnoreOnceQuickFix(state, new[] { inspection }).Fix(inspectionResults.First());
             Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
         }
+
+
+        [TestMethod]
+        [TestCategory("QuickFixes")]
+        public void ObsoleteCallStatement_IgnoreQuickFixWorks()
+        {
+            const string inputCode =
+@"Sub Foo()
+    Call Goo(1, ""test"")
+End Sub
+
+Sub Goo(arg1 As Integer, arg1 As String)
+    Call Foo
+End Sub";
+
+            const string expectedCode =
+@"Sub Foo()
+'@Ignore ObsoleteCallStatement
+    Call Goo(1, ""test"")
+End Sub
+
+Sub Goo(arg1 As Integer, arg1 As String)
+'@Ignore ObsoleteCallStatement
+    Call Foo
+End Sub";
+
+            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out var component);
+            var state = MockParser.CreateAndParse(vbe.Object);
+
+            var inspection = new ObsoleteCallStatementInspection(state);
+            var inspector = InspectionsHelper.GetInspector(inspection);
+            var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
+
+            var fix = new IgnoreOnceQuickFix(state, new[] { inspection });
+            foreach (var result in inspectionResults)
+            {
+                fix.Fix(result);
+            }
+
+            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
+        }
+
     }
 }
